minor fix: semantics returned value
Returned value is bool and not int. The value remains the same (0 = false) but this is clearer in the code
Also note LastError value isn't set to 0 when position at/beyond file. The caller may have to check or SetLastError(0) before using CascReadFile in a loop. Alternatively, this could done by CascReadFile in this case but this behavior is not changed for now.