Final changes to support custom CDN's
Instead of using the existing region, introduced CdnRegion, as this could differ to where ever the CDN is hosted., and should not effect picking a CDN host server.

Also added a second check to content-length headers, as HTTP standard says all header checks should be case insensitive, this was causing some issues with custom hosting.